add package.swift + apply swiftformat
