+++ serde-generate/runtime/swift/README.md
@@ -1,0 +1,3 @@
+# serde
+
+Provides bincode and BCS serialization runtimes for the data structures generated by `serde-generate`.
